Refactor Kotlin Emitter
