+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/emit/EmitSharedEmitter.kt
@@ -1,0 +1,5 @@
+package community.flock.wirespec.compiler.core.emit
+
+interface EmitSharedEmitter {
+    val emitShared: EmitShared
+}
