+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/emit/PackageNameEmitter.kt
@@ -1,0 +1,5 @@
+package community.flock.wirespec.compiler.core.emit
+
+interface PackageNameEmitter {
+    val packageName: PackageName
+}
